// front/SPOTFront/SPOTAPI/model/protagonist.ts
/**
 * SPOT API
 * API of SPOT Project
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Company } from './company';
import { Person } from './person';


export interface Protagonist { 
    id?: number;
    type: Protagonist.TypeEnum;
    name: string;
    link?: string;
    photo?: string;
    dateUpdate?: string;
    person?: Person;
    company?: Company;
}
export namespace Protagonist {
    export type TypeEnum = 'person' | 'company';
    export const TypeEnum = {
        Person: 'person' as TypeEnum,
        Company: 'company' as TypeEnum
    };
}